put transform buffer functions in cpp file
